Add page for when app links aren't resolved
When an Accrescent user presses an Android App Link, it opens the
corresponding app detail page in the client. However, if they don't have
Accrescent installed (or the link is not verifying for some reason, as
is the case for some devices now due to an AOSP bug), they were
previously faced with a 404.

This page explains why the user may have reached it (Accrescent isn't
installed or AOSP bug) and how to allow the Android App Link to work
(install Accrescent or add link manually respectively).