Improve open-source verbiage
Accrescent is open-source and has no plans to change this model.
However, open-source development is not the primary focus of Accrescent,
and giving that aspect of it too much weight could skew the public
perception of what Accrescent views as important. Thus, remove emphatic
statements such as "open-source forever" and "completely open-source" to
better align the website's content with Accrescent's stance.