Remove icon verification from website
Icon verification was never implemented fully and so really shouldn't
have been added to the site in the first place. We plan to remove
name/label verification as well (at least temporarily), but since it is
implemented currently, we're not going to remove it from the site yet.

The href is kept the same despite its reference to icon verification to
prevent external links from breaking.