Add back app size requirement to docs requirements section
We temporarily removed this requirement in dae75ab because we didn't
have a specific limit for app sizes. While we still don't have fully
fleshed-out requirements for maximum app sizes (see accrescent/meta#33
for that), we introduced a maximum APK set size of 1 GiB in
accrescent/parcelo@b39d1c4. Therefore, we should reintroduce the app
size requirement to the requirements section of our documentation to
notify developers.