Explicitly state we're not a registered nonprofit
We have never stated that we're a nonprofit, registered or otherwise. No
one has been confused by this so far, but to ensure that nobody gets
confused in the future and thinks we are a registered nonprofit, we
should add this disclaimer just to be safe and indicate that e.g.
donations are probably not tax-deductible depending on your tax laws.

This commit does not indicate any change in how we operate or use the
funds we receive. It only indicates that we are not legally considered a
nonprofit organization.